Removing ngrok
It's not you, it's err. Actually it is you.

failed: [localhost] => (item=ngrok) => {"failed": true, "item": "ngrok"}
msg: Error: No available formula with the name "ngrok"
Upstream sunsetted 1.x in March 2016 and 2.x is not open-source.

If you wish to use the 2.x release you can install it
either via Homebrew:
  brew install homebrew/binary/ngrok2

Or via the Cask:
  brew cask install ngrok
failed: [localhost] => (item=htop) => {"failed": true, "item": "htop"}
msg: Error: Cannot install htop because conflicting formulae are installed.

So we'll need to add this as a cask if we want it.